backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
httpx==0.25.2
python-dotenv==1.0.0
gitpython==3.1.40
google-generativeai>=0.8.0
